+++ src/main/java/jtp/cc1/monster/GridRC.java
@@ -1,4 +1,4 @@
-package jtp.cc1.ascsp1;
+package monster;
 
 import java.io.Serializable;
